# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://laravel-12.test"
        },
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "http://localhost:8000/api/v1"
        },
        {
            "url": "http://localhost:8000/api/v2"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Retrieve a paginated list of products with optional filtering and search",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by product status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "archived"
                            ]
                        }
                    },
                    {
                        "name": "vendor",
                        "in": "query",
                        "description": "Filter by vendor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_type",
                        "in": "query",
                        "description": "Filter by product type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "in_stock",
                        "in": "query",
                        "description": "Filter by stock availability",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in name, description, or SKU",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Create a new product with all required and optional fields",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product",
                "description": "Retrieve a single product by ID including its variants",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Update an existing product by ID",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Delete a product and all its variants",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "T-Shirt"
                    },
                    "description": {
                        "type": "string",
                        "example": "A comfortable cotton t-shirt"
                    },
                    "sku": {
                        "type": "string",
                        "example": "TSH-001"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "compare_price": {
                        "type": "number",
                        "format": "float",
                        "example": 39.99
                    },
                    "inventory_quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "track_inventory": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "archived"
                        ],
                        "example": "active"
                    },
                    "vendor": {
                        "type": "string",
                        "example": "Brand Name"
                    },
                    "product_type": {
                        "type": "string",
                        "example": "Clothing"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "cotton",
                            "casual"
                        ]
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "image1.jpg",
                            "image2.jpg"
                        ]
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 0.3
                    },
                    "weight_unit": {
                        "type": "string",
                        "enum": [
                            "kg",
                            "g",
                            "lb",
                            "oz"
                        ],
                        "example": "kg"
                    },
                    "requires_shipping": {
                        "type": "boolean",
                        "example": true
                    },
                    "seo": {
                        "properties": {
                            "title": {
                                "type": "string",
                                "example": "Comfortable Cotton T-Shirt"
                            },
                            "description": {
                                "type": "string",
                                "example": "Meta description for SEO"
                            },
                            "keywords": {
                                "type": "string",
                                "example": "t-shirt, cotton, clothing"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariant"
                        }
                    }
                },
                "type": "object"
            },
            "ProductVariant": {
                "title": "ProductVariant",
                "description": "Product variant model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Small / Red"
                    },
                    "sku": {
                        "type": "string",
                        "example": "TSH-001-S-RED"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "compare_price": {
                        "type": "number",
                        "format": "float",
                        "example": 39.99
                    },
                    "inventory_quantity": {
                        "type": "integer",
                        "example": 25
                    },
                    "track_inventory": {
                        "type": "boolean",
                        "example": true
                    },
                    "inventory_policy": {
                        "type": "string",
                        "enum": [
                            "deny",
                            "continue"
                        ],
                        "example": "deny"
                    },
                    "fulfillment_service": {
                        "type": "string",
                        "example": "manual"
                    },
                    "option1": {
                        "type": "string",
                        "example": "Small"
                    },
                    "option2": {
                        "type": "string",
                        "example": "Red"
                    },
                    "option3": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 0.3
                    },
                    "weight_unit": {
                        "type": "string",
                        "enum": [
                            "kg",
                            "g",
                            "lb",
                            "oz"
                        ],
                        "example": "kg"
                    },
                    "barcode": {
                        "type": "string",
                        "example": "1234567890123"
                    },
                    "image": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "variant1.jpg"
                        ]
                    },
                    "requires_shipping": {
                        "type": "boolean",
                        "example": true
                    },
                    "taxable": {
                        "type": "boolean",
                        "example": true
                    },
                    "position": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}